// db/supabase/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.4"
  }
  public: {
    Tables: {
// ...
      products: {
        Row: {
          approved: boolean
          categories: string | null
          codename: string
          created_at: string | null
          description: string
          email: string | null
          featured: boolean
          full_name: string | null
          id: string
          labels: string[] | null
          logo_src: string | null
          product_website: string
          punchline: string
          tags: string[] | null
          twitter_handle: string | null
          user_id: string
          view_count: number | null
        }
        Insert: {
          approved?: boolean
          categories?: string | null
          codename: string
          created_at?: string | null
          description: string
          email?: string | null
          featured?: boolean
          full_name?: string | null
          id?: string
          labels?: string[] | null
          logo_src?: string | null
          product_website: string
          punchline: string
          tags?: string[] | null
          twitter_handle?: string | null
          user_id: string
          view_count?: number | null
        }
        Update: {
          approved?: boolean
          categories?: string | null
          codename?: string
          created_at?: string | null
          description?: string
          email?: string | null
          featured?: boolean
          full_name?: string | null
          id?: string
          labels?: string[] | null
          logo_src?: string | null
          product_website?: string
          punchline?: string
          tags?: string[] | null
          twitter_handle?: string | null
          user_id?: string
          view_count?: number | null
        }
        Relationships: [
          {
            foreignKeyName: "products_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "user_admin_status"
            referencedColumns: ["id"]
          }
        ]
      }
// ...
    }
// ...
  }
}